frontend/requirements.txt
streamlit==1.35.0
httpx==0.27.0
python-dotenv==1.0.1
